liana-gui: move parse_args() to lib
