+++ liana-gui/src/main.rs
@@ -1,6 +1,6 @@
 #![windows_subsystem = "windows"]
 
-use std::{error::Error, fmt::Display, io::Write, path::PathBuf, process, str::FromStr};
+use std::{error::Error, io::Write, str::FromStr};
 
 #[cfg(target_os = "linux")]
 use iced::window::settings::PlatformSpecific;
@@ -18,81 +18,9 @@ use liana_gui::{
     dir::LianaDirectory,
     gui::{Config, LianaGUI},
     node::bitcoind::delete_all_bitcoind_locks_for_process,
+    utils::args::{parse_args, Arg},
     VERSION,
 };
-
-#[derive(Debug, Clone, PartialEq)]
-enum Arg {
-    DatadirPath(LianaDirectory),
-    Network(Network),
-}
-
-fn parse_args(
-    args: Vec<String>,
-    version: impl Display,
-    available_networks: &[Network],
-    default_network: Option<Network>,
-) -> Result<Vec<Arg>, Box<dyn Error>> {
-    let mut res = Vec::new();
-
-    let app_name = std::path::Path::new(&args[0])
-        .file_name()
-        .and_then(|s| s.to_str())
-        // This should never happen
-        .unwrap_or("liana");
-
-    if args.len() > 1 && (args[1] == "--version" || args[1] == "-v") {
-        eprintln!("{}", version);
-        process::exit(0);
-    }
-
-    if args.len() > 1 && (args[1] == "--help" || args[1] == "-h") {
-        let network_options: String = available_networks
-            .iter()
-            .map(|n| {
-                let name = n.to_string().to_lowercase();
-                let default_marker = if Some(*n) == default_network {
-                    " (default)"
-                } else {
-                    ""
-                };
-                format!("    --{:<15} Use {} network{}", name, name, default_marker)
-            })
-            .collect::<Vec<_>>()
-            .join("\n");
-
-        eprintln!(
-            r#"
-Usage: {app_name} [OPTIONS]
-
-Options:
-    --datadir <PATH>    Path of liana datadir
-    -v, --version       Display {app_name} version
-    -h, --help          Print help
-{network_options}
-        "#
-        );
-        process::exit(0);
-    }
-
-    for (i, arg) in args.iter().enumerate() {
-        if arg == "--datadir" {
-            if let Some(a) = args.get(i + 1) {
-                res.push(Arg::DatadirPath(LianaDirectory::new(PathBuf::from(a))));
-            } else {
-                return Err("missing arg to --datadir".into());
-            }
-        } else if arg.starts_with("--") && arg != "--datadir" {
-            let network = Network::from_str(arg.trim_start_matches("--"))?;
-            if !available_networks.contains(&network) {
-                return Err(format!("network {} is not available", network).into());
-            }
-            res.push(Arg::Network(network));
-        }
-    }
-
-    Ok(res)
-}
 
 fn main() -> Result<(), Box<dyn Error>> {
     use Network::{Bitcoin, Regtest, Signet, Testnet};
@@ -214,73 +142,4 @@ fn setup_panic_hook(liana_directory: &LianaDirectory) {
         std::io::stdout().flush().expect("Flushing stdout");
         std::process::exit(1);
     }));
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use liana_gui::dir::LianaDirectory;
-    use Network::{Bitcoin, Regtest, Signet, Testnet};
-
-    const ALL_NETWORKS: &[Network] = &[Bitcoin, Testnet, Signet, Regtest];
-
-    #[test]
-    fn test_parse_args() {
-        assert!(parse_args(
-            vec!["app".into(), "--meth".into()],
-            VERSION,
-            ALL_NETWORKS,
-            None
-        )
-        .is_err());
-        assert!(parse_args(
-            vec!["app".into(), "--datadir".into()],
-            VERSION,
-            ALL_NETWORKS,
-            None
-        )
-        .is_err());
-        assert_eq!(
-            Some(vec![Arg::Network(Regtest)]),
-            parse_args(
-                vec!["app".into(), "--regtest".into()],
-                VERSION,
-                ALL_NETWORKS,
-                None
-            )
-            .ok()
-        );
-        assert_eq!(
-            Some(vec![
-                Arg::DatadirPath(LianaDirectory::new(PathBuf::from("hello"))),
-                Arg::Network(Testnet)
-            ]),
-            parse_args(
-                "app --datadir hello --testnet"
-                    .split(' ')
-                    .map(|a| a.to_string())
-                    .collect(),
-                VERSION,
-                ALL_NETWORKS,
-                None
-            )
-            .ok()
-        );
-        assert_eq!(
-            Some(vec![
-                Arg::Network(Testnet),
-                Arg::DatadirPath(LianaDirectory::new(PathBuf::from("hello"))),
-            ]),
-            parse_args(
-                "app --testnet --datadir hello"
-                    .split(' ')
-                    .map(|a| a.to_string())
-                    .collect(),
-                VERSION,
-                ALL_NETWORKS,
-                None
-            )
-            .ok()
-        );
-    }
 }
