+++ liana-gui/src/utils/mod.rs
@@ -5,6 +5,7 @@ use std::{
 
 use liana::miniscript::bitcoin::{self, bip32::DerivationPath, Network};
 
+pub mod args;
 pub mod serde;
 
 #[cfg(test)]
